Created Decode Subsystem skeleton.
As of writing, no robot has been built- only designed. All code sonn to
be written nebulous, and most is treated abstractly. All that exists
physically are  disjoint prototypes (some uncompleted), and the plan is
that programmers have, in our 3 week schedule, a total of "You'll find a
way to work your magic" days.

¯\_(ツ)_/¯