Massively improved teleop auto reloading
This included the addition of distance sensors to
ArtifactColorRangeSensor, separate from the functionality of
ColorRangeSensors. This allows for better reliability in differentiating
between UNKNOWN and some known color (i.e. green or purple). This allows
us to filter out the colors better, making the tuning when the color is
within distance more recognizable.

Specifically in teleop, FlywheelTubeShooter had reloading separated into
two speeds: auto speed and manual speed. Auto is slower and lasts for a
shorter time. Manual is faster and lasts a little longer. Automatic
reloading in in teleop uses the slower version, as it's more reliable
for color sensing in time.