Created interpolation for ramp angles from distance
This is useful for converting a given distance measurement into an
angle of the ramp using experimental data. Not that the interpolation
between data points is linear. An improvement on this algorithm may
be to find the second derivative of the data and integrate twice so that
the resulting curve is smoother.

Food for thought. 🍗