Retuned the ramp angle using servo pos instead of inches
It was found that a distance sensor was noisy, would vary depending on
the day, and could be obstructed. A servo is more consistent even if
hardware changes could alter the angle. However, we can verify the
angle before a match using the level app and make sure the underside
angle of the ramp is 37 or 38 degrees.