Rewrote the shooter to maintain a provided charge speed, not the default
This allows a charge speed to be set by `charge(ins/sec)` and be
maintained even when firing or reloading. In contrast, the old version
would go back to whatever `FLYWHEEL_CONST.chargedSpeed` was if you fired
or reloaded. Using the no argument version of `charge()` does still use
this default, however.